backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.9
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.2.0
websockets==13.1
httpx==0.28.1
openai>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0